Add BCOV-ReactiveCocoa-iOS target for short term solution
This target will create a libBCOVRAC.a containing all of the ReactiveCocoa
symbols with the `BCOV_` namespacing prefix. It will  be added to the list of
libraries that the BCOVPlayerSDK target in BCOVPlayerSDK.xcodeproj links
against, for the short-term solution until we have a chance to incorporate
Xcode project-mangling to keep from having to maintain a modified clone of the
ReactiveCocoa repository.